fix(core): move formatDate to utils folder
