+++ apps/core/src/utils/index.ts
@@ -4,6 +4,7 @@
 export * from './calculateStakeShare';
 export * from './chunkArray';
 export * from './formatAmount';
+export * from './formatDate';
 export * from './formatPercentageDisplay';
 export * from './getRefGasPrice';
 export * from './hasDisplayData';
